+++ browser/modules/licenses.ts
@@ -12,20 +12,21 @@
  * permissions and limitations under the License.
  */
 
-import { WebLicense } from '../../server/api/licenses/interfaces';
+import { WebLicense, WebTag } from '../../server/api/licenses/interfaces';
 import { fetchAuth } from '../util';
 
 export const RECEIVE_LICENSES = 'app/licenses/receive-licenses';
-export const RECEIVE_CLAS = 'app/licenses/receive-clas';
 
 interface State {
   list: WebLicense[];
   map: Map<string, WebLicense>;
+  tags: {[key: string]: WebTag};
 }
 
 const initial: State = {
   list: [],
   map: new Map(),
+  tags: {},
 };
 
 export default function reducer(state = initial, action: any = {}): State {
@@ -38,6 +39,7 @@ export default function reducer(state = initial, action: any = {}): State {
           map.set(curr.name, curr);
           return map;
         }, state.map),
+        tags: action.tags,
       };
 
     default:
@@ -47,10 +49,11 @@ export default function reducer(state = initial, action: any = {}): State {
 
 /*** Action creators ***/
 
-export function receiveLicenses(licenses) {
+export function receiveLicenses(data) {
   return {
     type: RECEIVE_LICENSES,
-    licenses,
+    licenses: data.licenses,
+    tags: data.tags,
   };
 }