+++ server/api/licenses/index.ts
@@ -13,13 +13,29 @@
  */
 
 import { licenses, mapTag } from '../../licenses';
-import { LicensePresentation, WebLicense } from './interfaces';
+import { WebLicense, WebTag } from './interfaces';
+
+interface LicenseResponse {
+  licenses: WebLicense[];
+  tags: {[key: string]: WebTag};
+}
 
 let cachedLicenses: WebLicense[] = [];
+const cachedTags: {[key: string]: WebTag} = {};
 
-export async function listLicenses(): Promise<WebLicense[]> {
+export async function listLicenses(): Promise<LicenseResponse> {
   // this is _slightly_ expensive, so cache it
   if (cachedLicenses.length === 0) {
+    cacheLicenseData();
+  }
+
+  return {
+    licenses: cachedLicenses,
+    tags: cachedTags,
+  };
+}
+
+function cacheLicenseData() {
     // a little silly, but keep two license lists; one for those that ask to be
     // displayed first and the other for those that don't care
     const first = [];
@@ -27,14 +43,30 @@ export async function listLicenses(): Promise<WebLicense[]> {
 
     // load each license...
     for (const [id, data] of (licenses as any).entries()) {
-      // and their tag data
+      const tags = data.get('tags');
+
+      // fill the tag cache as we go
+      let sortFirst = false;
+      for (const tag of tags) {
+        const mod = mapTag(tag);
+        cachedTags[tag] = {
+          presentation: mod.presentation,
+          questions: mod.questions,
+        };
+
+        // check the sort option, since we handle that part server-side
+        if (mod.presentation && mod.presentation.sortFirst) {
+          sortFirst = true;
+        }
+      }
+
       const item = {
         name: id,
-        tags: data.get('tags'),
-        presentation: collectTagPresentations(data.get('tags')),
+        tags,
       };
+
       // send it to either of the lists by presentation preference
-      if (item.presentation && item.presentation.sortFirst) {
+      if (sortFirst) {
         first.push(item);
       } else {
         rest.push(item);
@@ -42,42 +74,4 @@ export async function listLicenses(): Promise<WebLicense[]> {
     }
 
     cachedLicenses = first.concat(rest);
-  }
-
-  return cachedLicenses;
-}
-
-/**
- * Mash individual tags' presentation fields into one when displaying a license.
- */
-function collectTagPresentations(tags: string[]): LicensePresentation | undefined {
-  const presentation: LicensePresentation = {};
-
-  for (const tag of tags) {
-    const mod = mapTag(tag);
-    const p = mod.presentation;
-    if (p) {
-      if (p.sortFirst) {
-        presentation.sortFirst = true;
-      }
-      if (p.fixedText) {
-        presentation.fixedText = true;
-      }
-      if (p.shortText) {
-        presentation.shortText = presentation.shortText || [];
-        presentation.shortText.push(p.shortText);
-      }
-      if (p.longText) {
-        presentation.longText = presentation.longText || [];
-        presentation.longText.push(p.longText);
-      }
-    }
-  }
-
-  // don't return anything if it's empty (no need to flood the JSON output, basically)
-  if (Object.keys(presentation).length === 0) {
-    return;
-  }
-
-  return presentation;
 }