+++ server/licenses/known/MyCustomLicense.ts
@@ -1,0 +1,7 @@
+export const tags = ['popular', 'linkage'];
+
+export const text = `
+This is a sample custom license.
+
+It's not very special, and it's also not really a license at all.
+`;